숙소 예약 배치프로그램 / 관리자페이지 송금신청내역

+++ src/main/java/com/hoge/vo/other/Host.java
@@ -21,7 +21,7 @@ public class Host {
 	private String name;
 	private int userNo;
 	private String accountHolderName;
-	private int accountNumber;
+	private long accountNumber;
 	private String bankName;
 	private int hostingType; // 1. 시설 2.체험
 	private String tel;
